Laurente new branch

+++ backend/codex/src/main/java/clinicaflow/controller/UserAccountController.java
@@ -4,15 +4,13 @@ import clinicaflow.entity.UserAccountEntity;
 import clinicaflow.service.UserAccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.bind.annotation.CrossOrigin;
-
 
 import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/useraccount")
-@CrossOrigin(origins = "http://localhost:3000") // Add this line
+@CrossOrigin(origins = "http://localhost:3000") 
 public class UserAccountController {
 
     @Autowired
@@ -32,19 +30,19 @@ public class UserAccountController {
 
     // READ ONE
     @GetMapping("/{id}")
-    public Optional<UserAccountEntity> getAccountById(@PathVariable Long id) {
+    public Optional<UserAccountEntity> getAccountById(@PathVariable int id) {
         return service.getAccountById(id);
     }
 
     // UPDATE
     @PutMapping("/update/{id}")
-    public UserAccountEntity updateAccount(@PathVariable Long id, @RequestBody UserAccountEntity account) {
+    public UserAccountEntity updateAccount(@PathVariable int id, @RequestBody UserAccountEntity account) {
         return service.updateAccount(id, account);
     }
 
     // DELETE
     @DeleteMapping("/delete/{id}")
-    public String deleteAccount(@PathVariable Long id) {
+    public String deleteAccount(@PathVariable int id) {
         service.deleteAccount(id);
         return "User account with ID " + id + " has been deleted.";
     }
