+++ backend/codex/src/main/java/clinicaflow/repository/UserAccountRepository.java
@@ -4,7 +4,7 @@ import clinicaflow.entity.UserAccountEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserAccountRepository extends JpaRepository<UserAccountEntity, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccountEntity, Integer> {
     Optional<UserAccountEntity> findByUsername(String username);
     boolean existsByUsername(String username); 
 }
